jsarcase/test/conformance2/rendering/blitframebuffer-stencil-only.html

<!--
Copyright (c) 2019 The Khronos Group Inc.
Use of this source code is governed by an MIT-style license that can be
found in the LICENSE.txt file.
-->

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>WebGL BlitFramebuffer Stencil-only Tests</title>
<link rel="stylesheet" href="../../resources/js-test-style.css"/>
<script src="../../js/js-test-pre.js"></script>
<script src="../../js/webgl-test-utils.js"></script>

<script id="vs" type="x-shader/x-vertex">#version 300 es
in vec4 position;
void main() {
  gl_Position = position;
}
</script>
<script id="fs" type="x-shader/x-fragment">#version 300 es
out mediump vec4 colorOut;
uniform mediump vec3 color;
void main() {
   colorOut = vec4(color, 1.0);
}
</script>

<script src="../../js/webgl-async-loader.js"></script>
</head>
<body>
<canvas id="example" width="8" height="8"></canvas>
<div id="description"></div>
<div id="console"></div>

<script type="module">
"use strict";
import * as loader from '../../js/webgl-async-loader.js';
const { waitForWebGLTestGlobals } = loader;

// 将原始测试代码封装为函数
function runWebGLTest() {
  try {


var wtu = WebGLTestUtils;
description("This test covers some edge cases of blitFramebuffer with stencil.");

var gl = wtu.create3DContext("example", undefined, 2);

var program, colorLoc;

function init_buffer(format) {
  var buf = gl.createFramebuffer();
  gl.bindFramebuffer(gl.FRAMEBUFFER, buf)
  var tex = gl.createTexture();
  gl.bindTexture(gl.TEXTURE_2D, tex);
  gl.texStorage2D(gl.TEXTURE_2D, 1, gl.RGBA8, 16, 16);
  gl.framebufferTexture2D(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, gl.TEXTURE_2D, tex, 0);
  var rbo = gl.createRenderbuffer();
  gl.bindRenderbuffer(gl.RENDERBUFFER, rbo);
  gl.renderbufferStorage(gl.RENDERBUFFER, format, 16, 16);
  gl.framebufferRenderbuffer(gl.FRAMEBUFFER,
      gl.DEPTH_STENCIL_ATTACHMENT, gl.RENDERBUFFER, rbo);

  gl.clearBufferfi(gl.DEPTH_STENCIL, 0, 1.0, 0);

  wtu.glErrorShouldBe(gl, gl.NO_ERROR, "after buffer init");
  shouldBe('gl.checkFramebufferStatus(gl.FRAMEBUFFER)', 'gl.FRAMEBUFFER_COMPLETE');

  return { fbo: buf, color: tex, depthStencil: rbo };
}

var quadVB;

function drawQuad(depth) {
  if (!quadVB) {
    quadVB = gl.createBuffer()
  }

  var quadVerts = new Float32Array(3 * 6);
  quadVerts[0] = -1.0; quadVerts[1] =   1.0; quadVerts[2] = depth;
  quadVerts[3] = -1.0; quadVerts[4] =  -1.0; quadVerts[5] = depth;
  quadVerts[6] =  1.0; quadVerts[7] =  -1.0; quadVerts[8] = depth;
  quadVerts[9] = -1.0; quadVerts[10] =  1.0; quadVerts[11] = depth;
  quadVerts[12] = 1.0; quadVerts[13] = -1.0; quadVerts[14] = depth;
  quadVerts[15] = 1.0; quadVerts[16] =  1.0; quadVerts[17] = depth;

  gl.bindBuffer(gl.ARRAY_BUFFER, quadVB);
  gl.bufferData(gl.ARRAY_BUFFER, quadVerts, gl.STATIC_DRAW);
  gl.vertexAttribPointer(0, 3, gl.FLOAT, gl.FALSE, 0, 0);
  gl.enableVertexAttribArray(0);
  gl.drawArrays(gl.TRIANGLES, 0, 6);

  wtu.glErrorShouldBe(gl, gl.NO_ERROR, "after drawQuad");
}

// Test based on dEQP-GLES3.functional.blit.depth_stencil.depth_24_stencil8_stencil_only
function test_stencil_only_blit(format) {
  debug("testing format: " + wtu.glEnumToString(gl, format))

  var src = init_buffer(format);
  var dest = init_buffer(format);

  gl.bindFramebuffer(gl.FRAMEBUFFER, src.fbo);
  gl.viewport(0, 0, 16, 16);

  // Fill source with red, depth = 0.5, stencil = 7
  gl.enable(gl.DEPTH_TEST);
  gl.enable(gl.STENCIL_TEST);
  gl.stencilOp(gl.KEEP, gl.KEEP, gl.REPLACE);
  gl.stencilFunc(gl.ALWAYS, 7, 0xFF);
  gl.uniform3f(colorLoc, 1.0, 0.0, 0.0);
  drawQuad(0.5);

  // Fill dest with yellow, depth = 0.0, stencil = 1
  gl.bindFramebuffer(gl.FRAMEBUFFER, dest.fbo);
  gl.stencilFunc(gl.ALWAYS, 1, 0xff);
  gl.uniform3f(colorLoc, 1.0, 1.0, 0.0);
  drawQuad(0.0);

  // Perform copy.
  gl.bindFramebuffer(gl.READ_FRAMEBUFFER, src.fbo);
  gl.bindFramebuffer(gl.DRAW_FRAMEBUFFER, dest.fbo);
  gl.blitFramebuffer(0, 0, 16, 16, 0, 0, 16, 16, gl.STENCIL_BUFFER_BIT, gl.NEAREST);

  // Render blue where depth < 0, decrement on depth failure.
  gl.bindFramebuffer(gl.FRAMEBUFFER, dest.fbo);
  gl.stencilOp(gl.KEEP, gl.DECR, gl.KEEP);
  gl.stencilFunc(gl.ALWAYS, 0, 0xff);

  gl.uniform3f(colorLoc, 0.0, 0.0, 1.0);
  drawQuad(0.0);

  // Render green where stencil == 6.
  gl.disable(gl.DEPTH_TEST);
  gl.stencilOp(gl.KEEP, gl.KEEP, gl.KEEP);
  gl.stencilFunc(gl.EQUAL, 6, 0xff);

  gl.uniform3f(colorLoc, 0.0, 1.0, 0.0);
  drawQuad(0.0);

  wtu.glErrorShouldBe(gl, gl.NO_ERROR, "after test");
  wtu.checkCanvasRect(gl, 0, 0, 16, 16, [0, 255, 0, 255],
                      "stencil test should be green");

  gl.deleteFramebuffer(src.fbo);
  gl.deleteFramebuffer(dest.fbo);
  gl.deleteTexture(src.color);
  gl.deleteTexture(dest.color);
  gl.deleteRenderbuffer(src.depthStencil);
  gl.deleteRenderbuffer(dest.depthStencil);
}

if (!gl) {
    testFailed("WebGL context does not exist");
} else {
    testPassed("WebGL context exists");

    program = wtu.setupProgram(gl, ["vs", "fs"], ["position"]);
    wtu.glErrorShouldBe(gl, gl.NO_ERROR, "after program initialization");
    shouldBe('gl.getProgramParameter(program, gl.LINK_STATUS)', 'true');

    colorLoc = gl.getUniformLocation(program, "color")
    wtu.glErrorShouldBe(gl, gl.NO_ERROR, "query uniform location");
    shouldBeNonNull('colorLoc')

    test_stencil_only_blit(gl.DEPTH24_STENCIL8);
    test_stencil_only_blit(gl.DEPTH32F_STENCIL8);
}


    var successfullyParsed = true;
    console.log("✅ WebGL test completed successfully!");
  } catch (error) {
    console.error("❌ WebGL test failed:", error);
    if (typeof debug !== 'undefined') {
      debug("Test failed: " + error.message);
    }
  }
}

// 环境检测函数
function isJSARRuntime() {
  const checks = [
    () => typeof waitForWebGLTestGlobals !== 'undefined',
    () => typeof process !== 'undefined' && process.versions && process.versions.node,
    () => typeof navigator !== 'undefined' && navigator.userAgent && navigator.userAgent.includes('JSAR'),
    () => typeof global !== 'undefined' && global.process,
    () => typeof window === 'undefined' && typeof global !== 'undefined',
    () => typeof module !== 'undefined' && module.exports
  ];
  
  const result = checks.some(check => {
    try {
      return check();
    } catch (e) {
      return false;
    }
  });
  
  console.log('🔍 JSAR Runtime detection result:', result);
  return result;
}

// 简单轮询等待函数
function waitForGlobalsSimple(globals, callback, maxAttempts = 300) {
  let attempts = 0;
  console.log('🔄 Starting simple polling for globals:', globals);
  
  function check() {
    attempts++;
    const available = [];
    const missing = [];
    
    globals.forEach(name => {
      try {
        if (typeof eval(name) !== 'undefined') {
          available.push(name);
        } else {
          missing.push(name);
        }
      } catch (e) {
        missing.push(name);
      }
    });
    
    if (missing.length === 0) {
      console.log('✅ All globals available:', globals);
      callback();
    } else if (attempts >= maxAttempts) {
      console.error('❌ Timeout waiting for globals. Available:', available, 'Missing:', missing);
    } else {
      if (attempts % 20 === 0 || attempts <= 5) {
        console.log(`⏳ Waiting for globals... Available: [${available.join(', ')}], Missing: [${missing.join(', ')}] (${attempts}/${maxAttempts})`);
      }
      setTimeout(check, 50);
    }
  }
  
  check();
}

// 智能初始化
function initializeTest() {
  console.log("🔍 Checking environment and dependencies...");
  
  const isJSAR = isJSARRuntime();
  const hasAsyncLoader = typeof waitForWebGLTestGlobals !== 'undefined';
  const requiredGlobals = ['description', 'debug', 'WebGLTestUtils'];
  
  console.log("Environment info:", {
    isJSARRuntime: isJSAR,
    hasAsyncLoader: hasAsyncLoader,
    userAgent: typeof navigator !== 'undefined' ? navigator.userAgent : 'N/A'
  });
  
  if (isJSAR || hasAsyncLoader) {
    console.log("🔄 JSAR Runtime environment detected, using async loader...");
    
    if (hasAsyncLoader) {
      waitForWebGLTestGlobals(runWebGLTest, {
        maxRetries: 300,
        retryInterval: 50,
        timeout: 15000,
        onTimeout: function() {
          console.error("❌ Timeout: WebGL test dependencies not available after 15s");
          try {
            if (typeof debug !== 'undefined') {
              debug("Test failed: Timeout waiting for dependencies");
            }
          } catch (e) {
            console.error("Debug function not available:", e);
          }
        },
        onError: function() {
          console.error("❌ Error: Failed to load WebGL test dependencies");
          try {
            if (typeof debug !== 'undefined') {
              debug("Test failed: Failed to load dependencies");
            }
          } catch (e) {
            console.error("Debug function not available:", e);
          }
        }
      });
    } else {
      console.log("⏳ No async loader found, using simple polling...");
      waitForGlobalsSimple(requiredGlobals, runWebGLTest);
    }
  } else {
    console.log("🌐 Browser environment detected, running test directly...");
    setTimeout(runWebGLTest, 500);
  }
}

// 确保DOM加载完成后再初始化
setTimeout(initializeTest, 500);

</script>
<!-- <script src="../../js/js-test-post.js"></script> -->

</body>
</html>
